-warning

+++ rts/Rendering/GL/VBO.cpp
@@ -176,7 +176,7 @@ bool VBO::BindBufferRangeImpl(GLenum target, GLuint index, GLuint _vboId, GLuint
 
 	const size_t neededAlignment = GetOffsetAlignment(target);
 	if (offset % neededAlignment != 0 || size % neededAlignment != 0) { //assert(?)
-		LOG_L(L_ERROR, "[VBO::%s]: attempt to bind with wrong offset [%u] or size [%I64u]. Needed alignment [%I64u]", __func__, offset, size, neededAlignment);
+		LOG_L(L_ERROR, "[VBO::%s]: attempt to bind with wrong offset [%I64u] or size [%I64u]. Needed alignment [%I64u]", __func__, static_cast<size_t>(offset), size, neededAlignment);
 		return false;
 	}
